+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Improved
 * game overview: improve how editions and purchases are displayed
 * add purchase: only allow choosing purchases of selected edition
+* session list: clicking the "End now?" link is not much faster
 
 ## 1.5.1 / 2023-11-14 21:10+01:00